replace: app router(archive page) -> tags/[slug]
