+++ src/app/(ArchivePage)/tags/[slug]/libs/index.ts
@@ -1,2 +1,1 @@
 export { getStaticPathsTagDetail } from './getStaticPathsTagDetail';
-export { getStaticPropsTagDetail } from './getStaticPropsTagDetail';
